[OpenMP] Replace math headers with OpenMP wrapper calls
This patch replaces the math functions when called on the device with an
`__omp_` variant that we can use. This will later be turned back into
the regular math call so we can avoid the problems with including math.h
on the device. This patch introduces the `-fopenmp-device-libm` flag
which enables the new math wrappers by defining __MATH_WRAPPERS__ in the
header.

Differential Revision: https://reviews.llvm.org/D121466